+++ contracts/factory/src/tests.rs
@@ -1,6 +1,8 @@
 use crate::contract::{Factory, FactoryClient};
 use soroban_sdk::{testutils::Address as _, Address, Env};
 
+use self::setup::install_multihop_wasm;
+
 mod config;
 mod setup;
 
@@ -15,6 +17,8 @@ fn test_deploy_factory_twice_should_fail() {
     let admin = Address::random(&env);
 
     let multihop = FactoryClient::new(&env, &env.register_contract(None, Factory {}));
-    multihop.initialize(&admin);
-    multihop.initialize(&admin);
+    let multihop_wasm_hash = install_multihop_wasm(&env);
+
+    multihop.initialize(&admin, &multihop_wasm_hash);
+    multihop.initialize(&admin, &multihop_wasm_hash);
 }