+++ contracts/factory/src/tests/setup.rs
@@ -1,12 +1,18 @@
 use crate::contract::{Factory, FactoryClient};
 use soroban_sdk::testutils::Address as _;
 use soroban_sdk::{Address, BytesN, Env};
-
 #[allow(clippy::too_many_arguments)]
 pub mod lp_contract {
     soroban_sdk::contractimport!(
         file = "../../target/wasm32-unknown-unknown/release/phoenix_pool.wasm"
     );
+}
+
+pub fn install_multihop_wasm(env: &Env) -> BytesN<32> {
+    soroban_sdk::contractimport!(
+        file = "../../target/wasm32-unknown-unknown/release/phoenix_multihop.wasm"
+    );
+    env.deployer().upload_contract_wasm(WASM)
 }
 
 pub fn install_lp_contract(env: &Env) -> BytesN<32> {
@@ -33,7 +39,8 @@ pub fn deploy_factory_contract<'a>(
 ) -> FactoryClient<'a> {
     let admin = admin.into().unwrap_or(Address::random(env));
     let factory = FactoryClient::new(env, &env.register_contract(None, Factory {}));
+    let multihop_wasm_hash = install_multihop_wasm(env);
 
-    factory.initialize(&admin);
+    factory.initialize(&admin, &multihop_wasm_hash);
     factory
 }